betty done

+++ _strlen.c
@@ -4,13 +4,13 @@
  *@s : string
  *Return: string length
  */
-size_t _strlen(const char *s)
+int _strlen(const char *s)
 {
-size_t i = 0;
+	int i = 0;
 
-for (i = 0; s[i] != '\0'; i++)
-{
+	for (i = 0; s[i] != '\0'; i++)
+	{
 
-}
-return (i);
+	}
+	return (i);
 }
